+++ src/auto-reply/reply/stranded-reply-recovery.ts
@@ -1,4 +1,4 @@
-import { formatSystemTurnPrompt } from "../../sessions/input-provenance.js";
+import { formatSystemTurnPrompt } from "../../sessions/system-turn-prompt.js";
 import type { SourceReplyDeliveryMode } from "../get-reply-options.types.js";
 import { markReplyPayloadForSourceSuppressionDelivery } from "../reply-payload.js";
 import type { ReplyPayload } from "../types.js";
